+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [1.3.1] - 2025-08-18
 
 ### Added
-- Open in window link
 - Added drag-and-drop/hide navigation customization
+- Match system theme toggle
+- Open in window link
+
 
 ## [1.3.0] - 2025-08-13